Standardize breach probability metric
